moving classes into corresponding packages
