+++ src/main/java/com/alopez/store/payments/services/PaymentGateway.java
@@ -1,6 +1,9 @@
 package com.alopez.store.payments.services;
 
 import com.alopez.store.orders.entities.Order;
+import com.alopez.store.payments.dtos.CheckoutSession;
+import com.alopez.store.payments.dtos.PaymentResult;
+import com.alopez.store.payments.dtos.WebhookRequest;
 
 import java.util.Optional;
